% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_multi_directional_derivative.R
\name{calc_multi_directional_derivative}
\alias{calc_multi_directional_derivative}
\title{Weighted multi-objective directional derivative}
\usage{
calc_multi_directional_derivative(dd_list, eta)
}
\arguments{
\item{dd_list}{A named list of directional derivative vectors.}

\item{eta}{A named numeric vector of weights.}
}
\value{
A numeric vector of combined directional derivative values.
}
\description{
Weighted multi-objective directional derivative
}
