% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_huang2020.R
\name{gt_huang2020_cost}
\alias{gt_huang2020_cost}
\title{Standardized cost for pool size (Huang et al. 2020)}
\usage{
gt_huang2020_cost(x, q)
}
\arguments{
\item{x}{Pool size (positive integer).}

\item{q}{Cost ratio in \verb{[0, 1]}.}
}
\value{
Scalar cost.
}
\description{
\deqn{c(x) = 1 - q + q x}
with \eqn{q = q_1 / (q_0 + q_1)} for assay cost \eqn{q_0} and enrollment cost
\eqn{q_1}. Value \eqn{q = 0} gives \eqn{c(x) \equiv 1}.
}
