% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_huang2020.R
\name{gt_huang2020_lambda}
\alias{gt_huang2020_lambda}
\title{Weight \eqn{\lambda(x)} in the information matrix (Huang et al. 2020)}
\usage{
gt_huang2020_lambda(x, theta, q)
}
\arguments{
\item{x}{Pool size.}

\item{theta}{Same as for \code{\link{gt_huang2020_pi}}.}

\item{q}{Same as for \code{\link{gt_huang2020_cost}}.}
}
\value{
Scalar (positive when \eqn{\pi} is interior).
}
\description{
\deqn{\lambda(x) = [c(x)\,\pi(x)\{1-\pi(x)\}]^{-1}}
}
