% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_equivalence.R
\name{plot_equivalence}
\alias{plot_equivalence}
\title{Plot equivalence theorem directional derivative}
\usage{
plot_equivalence(
  eq_obj,
  main = NULL,
  xlab = "Design point",
  ylab = "Directional derivative",
  pch_support = 19,
  ...
)
}
\arguments{
\item{eq_obj}{Output from \code{\link[=check_equivalence]{check_equivalence()}}.}

\item{main}{Plot title.}

\item{xlab}{X-axis label.}

\item{ylab}{Y-axis label.}

\item{pch_support}{Plotting symbol for support points.}

\item{...}{Passed to \code{\link[graphics:plot]{graphics::plot()}}.}
}
\value{
Invisibly returns \code{eq_obj}.
}
\description{
Plot equivalence theorem directional derivative
}
